fix(ontology): add property to res class
